# docs/openapi.yaml
openapi: 3.1.0
...
info:
  title: Angui API
  version: 0.1.0
  summary: Angui 搜救协作服务的认证、案件成员、案件和线索 API
  description: |
    本文档描述当前 Rust/Actix Web 服务实际注册的 HTTP API。

    除健康检查和登录外，所有接口都要求使用服务端会话 Bearer token。
    案件访问还取决于当前用户是否属于目标案件以及其案件角色。

    重要授权规则：

    - `member` 可以创建案件；创建者会成为该案件的 `family` 成员。
    - `commander` 与 `volunteer` 是可叠加的全局能力，不会自动产生案件成员关系。
    - family 只能显式邀请持有 commander 能力的活动成员；commander 可以添加 family、commander 或 volunteer，但同名案件角色要求目标成员持有相应全局能力。
    - commander 可以查看和审核案件内全部线索，并改变案件状态。
    - family 可以查看已确认线索及本人提交的线索。
    - volunteer 只能查看已确认线索，且 elder_profile.health_notes 会由服务端返回为 null。
    - 没有案件成员关系时，案件资源通常返回 404，以减少资源 ID 枚举。

    阅读约定：

    - 每个 operation 的 security 明确标出是否需要 Bearer 会话；没有 security 的公共接口会显式使用空数组。
    - x-account-types 表示可调用该 operation 的长期账号类型；x-case-roles 表示还必须具备的案件成员角色。全局能力仅用于成员分配资格，不能自动授予案件访问权。
    - x-data-classification 标出响应中需要按案件成员关系和角色裁剪的数据。客户端不得把 case-restricted 或 sensitive 字段缓存到公共页面、日志或分析服务。

    当前仍是 MVP API，不应处理真实搜救数据。正式部署必须使用 HTTPS。
  contact:
    name: AG-Angui
    url: https://github.com/AG-Angui/Angui
servers:
  - url: http://127.0.0.1:8080
    description: 本地 Actix Web 开发服务器
...
security:
  - bearerAuth: []
paths:
...
  /api/learning/resources:
    get:
      tags: [Learning]
      operationId: listLearningResources
      summary: List reviewed learning resources visible to the current account
      description: Returns only published, effective resources whose visibility matches the authenticated account. Withdrawn, future, internal-case, and raw archive material are never returned.
      security: [{ bearerAuth: [] }]
      x-account-types: [member, learner]
      x-data-classification: public-or-training
      responses:
        "200": { description: Reviewed learning resources }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/learning/questions:
    get:
      tags: [Learning]
      operationId: listLearningQuestions
      summary: List visible theory questions without answers
      description: Correct options, explanations, and unpublished questions are excluded. Filter by optional tag or difficulty.
      security: [{ bearerAuth: [] }]
      x-account-types: [member, learner]
      x-data-classification: training
      responses:
        "200": { description: Question prompts and options only }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/learning/questions/{question_id}/answers:
    post:
      tags: [Learning]
      operationId: submitLearningQuestionAnswer
      summary: Submit one option and receive a source-backed explanation
      description: Server-side validation checks visibility, published version, and that the submitted option belongs to the requested question. The result does not expose answers for other questions.
      security: [{ bearerAuth: [] }]
      x-account-types: [member, learner]
      x-data-classification: training
      responses:
        "200": { description: Correctness and source-backed explanation }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/knowledge/ask:
    post:
      tags: [Learning]
      operationId: askKnowledge
      summary: Answer using only visible reviewed learning resources
      description: Uses a deterministic source-matching fallback. When no reliable source is visible it returns an explicit insufficient-sources response and an instruction to seek human verification.
      security: [{ bearerAuth: [] }]
      x-account-types: [member, learner]
      x-data-classification: training
      responses:
        "200": { description: Source-backed or insufficient-sources answer }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Angui session token
      description: 登录接口返回的 angui_ 前缀服务端会话令牌。
...
  responses:
    ValidationError:
      description: 请求字段、枚举值或业务输入无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: validation_error
              message: last_seen_location is required
    Unauthorized:
      description: 缺少会话、Authorization 头无效、会话已过期或已撤销
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: unauthorized
              message: invalid or expired session
...
    NotFound:
      description: 资源不存在，或当前用户没有相应案件成员关系
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: not_found
              message: case was not found
...
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - validation_error
            - unauthorized
            - forbidden
            - not_found
            - conflict
            - rate_limited
            - internal_error
            - database_error
        message:
          type: string